Update client_certificate documentation in admin API Lua file
This commit updates the documentation for the `client_certificate` object in the admin API Lua file. The existing documentation suggested that an object should be sent with an `id` key, which caused confusion and inconsistencies when using the declarative configuration. However, through issue #8881 (link: https://github.com/Kong/kong/issues/8881), it was discovered that the `id` attribute is not required or supported in the declarative configuration. 

This update replaces the existing description and code snippets related to the `client_certificate` object with an updated description that reflects the correct behavior. The new description clarifies that if set, the `client_certificate` field specifies the ID of the client certificate to be used during TLS handshaking with the upstream server.

By making this change, we ensure that the documentation accurately represents the behavior of the `client_certificate` field and provides clear instructions to users. This resolves the issue reported by @meesvandongen on May 30, 2022, as documented in issue #8881.

Link to the issue: https://github.com/Kong/kong/issues/8881